requirements.txt
beautifulsoup4==4.13.3
colorama==0.4.6
cursor==1.3.5
ipwhois==1.3.0
matplotlib==3.10.1
pan_python==0.17.0
pyperclip==1.9.0
Requests==2.32.3
tabulate==0.9.0
urllib3==2.3.0
xmltodict==0.12.0
